requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-docx==1.1.2
python-pptx==0.6.23
lxml==5.4.0
httpx==0.27.0
azure-identity==1.16.0
python-multipart==0.0.9
